Remember what files have been processed
Fixes #35

gulp-remember is useful if the templates are preprocessed by e.g jade.

As gulp-angular-templatecache is a process + concat, plugin, this gives issues.

gulp-remember caches the whole file object, but templateCache is overwritting the contents of that object.
When the stream re-runs without change to some files, the Buffer output by templateCache is resent as Input, which is not what we want.

Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>