Fix removal of the '.' when used as the root
Sometimes you actually need the `.` prefix. Node's `path.join` drops the
`.` because it has no effect as far as the file system directory structure
is concerned.

One particular case for needing the '.' is adherence to naming conventions
of the new Angular router, which expects components to appear under a directory
structure similar to `./components/my-component/my-component.html`.