Specified empty dependencies in templates angular.module
I spent an hour debugging this and eventually I just added an empty dependency as the second argument and everything worked.

Error that was occurring:

```
'app.components.templates' is not available! You either misspelled the module name or forgot to load it. If registering a module ensure that you specify the dependencies as the second argument.
```